+++ R/app-ui.R
@@ -1,72 +1,26 @@
-#' App UI
+#' @title App UI
 #'
-#' Create the UI component of the dccvalidator Shiny app.
+#' @description If not running interactively (i.e. locally), launches the
+#' \code{\link{mod_synapse_oauth_ui}} to start dccvalidator using OAuth for
+#' login. Otherwise, launches the \code{\link{mod_main_ui}} to start
+#' dccvalidator using login credentials stored in a .synapseConfig.
 #'
+#' @export
 #' @import shiny
 #' @import shinydashboard
-#' @param request Shiny request
+#' @param request Shiny request object
 #' @return A shinydashboard page
-#' @export
 #' @examples
 #' \dontrun{
 #' shinyApp(ui = app_ui, server = app_server)
 #' }
 app_ui <- function(request) {
-  dashboardPage(
-    dashboardHeader(title = "Metadata Validation"),
-    dashboardSidebar(
-      sidebarMenu(
-        if (!is.na(config::get("path_to_markdown"))) {
-          menuItem("Using the App", tabName = "vignette")
-        },
-        if (config::get("docs_tab")$include_tab) {
-          menuItem(
-            config::get("docs_tab")$tab_name,
-            tabName = "documentation") 
-        },
-        menuItem("Validator", tabName = "validator")
-      ),
-      create_footer(config::get("contact_email"))
-    ),
-    dashboardBody(
-
-      # Add resources in www
-      golem_add_external_resources(),
-
-      # Use shinyjs
-      shinyjs::useShinyjs(),
-
-      # Make a list of the tabItems; this is a workaround
-      # for a problem with tabItems and shinyDashboard
-      tags$div(
-        list(
-          # Embedd How To Use App vignette
-          if (!is.na(config::get("path_to_markdown"))) {
-            tabItem(
-              tabName = "vignette",
-              get_markdown(config::get("path_to_markdown"))
-            )
-          },
-          # Validator UI
-          validator_ui(
-            id = "validator",
-            species_list = config::get("species_list"),
-            assay_templates = config::get("templates")$assay_templates
-          ),
-          # Documentation tab UI
-          if (config::get("docs_tab")$include_tab) {
-            # Documentation tab UI
-            upload_documents_ui(
-              id = "documentation",
-              markdown_path = config::get("docs_tab")$path_to_docs_markdown,
-              include_widget = config::get("docs_tab")$include_upload_widget
-            )
-          }
-        ),
-        class = "tab-content"
-      )
-    )
-  )
+  if (interactive()) {
+    ## Running locally; skip OAuth
+    mod_main_ui("main")
+  } else {
+    mod_synapse_oauth_ui(id = "oauth", request = request)
+  }
 }
 
 #' @import shiny